UI tweaks using Coordinator Layout during scroll - in friends, channels. UI tweaks to comply to material design guidelines regarding making elements more lifelike. Improvement of alarm activation screen, bigger image, etc. Inclusion of Node API for friend retrieval using boolean values. Better caching of friends in listview. Correction of a number of file deletion problems - TODO: channel download creating multiple copies to due to function error in manager class. On alarm enable, show time set.

+++ app/src/main/java/com/roostermornings/android/activity/NewAudioRecordActivity.java
@@ -135,6 +135,7 @@ public class NewAudioRecordActivity extends BaseActivity {
         super.onDestroy();
         if(!fileProcessed) {
             try {
+                if(mAudioSavePathInDevice == null) return;
                 final File localFile = new File(mAudioSavePathInDevice);
                 localFile.delete();
             } catch(Exception e) {
@@ -461,6 +462,8 @@ public class NewAudioRecordActivity extends BaseActivity {
                             PackageManager.PERMISSION_GRANTED;
 
                     if (StoragePermission && RecordPermission) {
+                        setNewAudioStatus(NEW_AUDIO_READY_LISTEN);
+                        startStopAudioRecording();
                     } else {
                         Toast.makeText(NewAudioRecordActivity.this, "Permission Denied", Toast.LENGTH_LONG).show();
                     }
